requirements.txt
openai
singlestoredb
SQLAlchemy
pymysql
pandas
dash
flask
plotly
python-dotenv
tenacity
